Create simple HAL for the bootloader
This is nothing fancy, and it mostly just consists of static inline
functions that represent tasks that a bootloader on any hardware would
need to accomplish.

This prepares us for potentially porting the SIMM programmer to a new
hardware architecture. The main bootloader algorithm can still be reused
on a new architecture; you would just need to provide any
hardware-specific stuff in a new hardware.c and hardware.h.
