+++ hal/at90usb646/LUFAConfig.h
